+++ src/Simplify.Mail/IMailSender.cs
@@ -66,9 +66,6 @@ public interface IMailSender : IDisposable
 	/// <param name="body">e-mail body.</param>
 	/// <param name="bodyForAntiSpam">Part of an e-mail body just for anti-spam checking.</param>
 	/// <param name="attachments">The attachments to an e-mail.</param>
-	/// <returns>
-	/// Process status, <see langword="true" /> if message is processed to sent successfully
-	/// </returns>
 	void Send(SmtpClient client, string from, string to, string subject, string body, string bodyForAntiSpam = null,
 		params MimeEntity[] attachments);
 
@@ -83,9 +80,6 @@ public interface IMailSender : IDisposable
 	/// <param name="bodyForAntiSpam">Part of an e-mail body just for anti-spam checking.</param>
 	/// <param name="attachments">The attachments to an e-mail.</param>
 	/// <param name="cancellationToken">The cancellation token.</param>
-	/// <returns>
-	/// Process status, <see langword="true" /> if message is processed to sent successfully
-	/// </returns>
 	Task SendAsync(SmtpClient client, string from, string to, string subject, string body, string bodyForAntiSpam = null,
 		MimeEntity[] attachments = null, CancellationToken cancellationToken = default);
 
@@ -98,7 +92,6 @@ public interface IMailSender : IDisposable
 	/// <param name="body">e-mail body.</param>
 	/// <param name="bodyForAntiSpam">Part of an e-mail body just for anti-spam checking.</param>
 	/// <param name="attachments">The attachments to an e-mail.</param>
-	/// <returns>Process status, <see langword="true"/> if message is processed to sent successfully</returns>
 	void Send(string from, string to, string subject, string body, string bodyForAntiSpam = null,
 		params MimeEntity[] attachments);
 
@@ -112,7 +105,6 @@ public interface IMailSender : IDisposable
 	/// <param name="bodyForAntiSpam">Part of an e-mail body just for anti-spam checking.</param>
 	/// <param name="attachments">The attachments to an e-mail.</param>
 	/// <param name="cancellationToken">The cancellation token.</param>
-	/// <returns>Process status, <see langword="true"/> if message is processed to sent successfully</returns>
 	Task SendAsync(string from, string to, string subject, string body, string bodyForAntiSpam = null,
 		MimeEntity[] attachments = null, CancellationToken cancellationToken = default);
 
@@ -125,10 +117,10 @@ public interface IMailSender : IDisposable
 	/// <param name="subject">e-mail subject.</param>
 	/// <param name="body">e-mail body.</param>
 	/// <param name="bodyForAntiSpam">Part of an e-mail body just for anti-spam checking.</param>
+	/// <param name="bccAddresses">BCC recipients.</param>
 	/// <param name="attachments">The attachments to an e-mail.</param>
-	/// <returns>Process status, <see langword="true"/> if all messages are processed to sent successfully</returns>
 	void Send(SmtpClient client, string fromMailAddress, IList<string> addresses, string subject, string body,
-		string bodyForAntiSpam = null, params MimeEntity[] attachments);
+		string bodyForAntiSpam = null, IList<string> bccAddresses = null, params MimeEntity[] attachments);
 
 	/// <summary>
 	/// Send e-mail to multiple recipients in one e-mail asynchronously.
@@ -139,11 +131,11 @@ public interface IMailSender : IDisposable
 	/// <param name="subject">e-mail subject.</param>
 	/// <param name="body">e-mail body.</param>
 	/// <param name="bodyForAntiSpam">Part of an e-mail body just for anti-spam checking.</param>
+	/// <param name="bccAddresses">BCC recipients.</param>
 	/// <param name="attachments">The attachments to an e-mail.</param>
 	/// <param name="cancellationToken">The cancellation token.</param>
-	/// <returns>Process status, <see langword="true"/> if all messages are processed to sent successfully</returns>
 	Task SendAsync(SmtpClient client, string fromMailAddress, IList<string> addresses, string subject, string body,
-		string bodyForAntiSpam = null, MimeEntity[] attachments = null, CancellationToken cancellationToken = default);
+		string bodyForAntiSpam = null, IList<string> bccAddresses = null, MimeEntity[] attachments = null, CancellationToken cancellationToken = default);
 
 	/// <summary>
 	/// Send e-mail to multiple recipients in one e-mail.
@@ -153,10 +145,10 @@ public interface IMailSender : IDisposable
 	/// <param name="subject">e-mail subject.</param>
 	/// <param name="body">e-mail body.</param>
 	/// <param name="bodyForAntiSpam">Part of an e-mail body just for anti-spam checking.</param>
+	/// <param name="bccAddresses">BCC recipients.</param>
 	/// <param name="attachments">The attachments to an e-mail.</param>
-	/// <returns>Process status, <see langword="true"/> if all messages are processed to sent successfully</returns>
 	void Send(string fromMailAddress, IList<string> addresses, string subject, string body,
-		string bodyForAntiSpam = null, params MimeEntity[] attachments);
+		string bodyForAntiSpam = null, IList<string> bccAddresses = null, params MimeEntity[] attachments);
 
 	/// <summary>
 	/// Send e-mail to multiple recipients in one e-mail asynchronously.
@@ -166,11 +158,11 @@ public interface IMailSender : IDisposable
 	/// <param name="subject">e-mail subject.</param>
 	/// <param name="body">e-mail body.</param>
 	/// <param name="bodyForAntiSpam">Part of an e-mail body just for anti-spam checking.</param>
+	/// <param name="bccAddresses">BCC recipients.</param>
 	/// <param name="attachments">The attachments to an e-mail.</param>
 	/// <param name="cancellationToken">The cancellation token.</param>
-	/// <returns>Process status, <see langword="true"/> if all messages are processed to sent successfully</returns>
 	Task SendAsync(string fromMailAddress, IList<string> addresses, string subject, string body,
-		string bodyForAntiSpam = null, MimeEntity[] attachments = null, CancellationToken cancellationToken = default);
+		string bodyForAntiSpam = null, IList<string> bccAddresses = null, MimeEntity[] attachments = null, CancellationToken cancellationToken = default);
 
 	/// <summary>
 	/// Send e-mail to multiple recipients and carbon copy recipients in one e-mail.
@@ -182,10 +174,10 @@ public interface IMailSender : IDisposable
 	/// <param name="subject">e-mail subject.</param>
 	/// <param name="body">e-mail body.</param>
 	/// <param name="bodyForAntiSpam">Part of an e-mail body just for anti-spam checking.</param>
+	/// <param name="bccAddresses">BCC recipients.</param>
 	/// <param name="attachments">The attachments to an e-mail.</param>
-	/// <returns>Process status, <see langword="true"/> if all messages are processed to sent successfully</returns>
 	void Send(SmtpClient client, string fromMailAddress, IList<string> addresses, IList<string> ccAddresses,
-		string subject, string body, string bodyForAntiSpam = null, params MimeEntity[] attachments);
+		string subject, string body, string bodyForAntiSpam = null, IList<string> bccAddresses = null, params MimeEntity[] attachments);
 
 	/// <summary>
 	/// Send e-mail to multiple recipients and carbon copy recipients in one e-mail asynchronously.
@@ -197,11 +189,11 @@ public interface IMailSender : IDisposable
 	/// <param name="subject">e-mail subject.</param>
 	/// <param name="body">e-mail body.</param>
 	/// <param name="bodyForAntiSpam">Part of an e-mail body just for anti-spam checking.</param>
+	/// <param name="bccAddresses">BCC recipients.</param>
 	/// <param name="attachments">The attachments to an e-mail.</param>
 	/// <param name="cancellationToken">The cancellation token.</param>
-	/// <returns>Process status, <see langword="true"/> if all messages are processed to sent successfully</returns>
 	Task SendAsync(SmtpClient client, string fromMailAddress, IList<string> addresses, IList<string> ccAddresses,
-		string subject, string body, string bodyForAntiSpam = null, MimeEntity[] attachments = null, CancellationToken cancellationToken = default);
+		string subject, string body, string bodyForAntiSpam = null, IList<string> bccAddresses = null, MimeEntity[] attachments = null, CancellationToken cancellationToken = default);
 
 	/// <summary>
 	/// Send e-mail to multiple recipients and carbon copy recipients in one e-mail.
@@ -212,10 +204,10 @@ public interface IMailSender : IDisposable
 	/// <param name="subject">e-mail subject.</param>
 	/// <param name="body">e-mail body.</param>
 	/// <param name="bodyForAntiSpam">Part of an e-mail body just for anti-spam checking.</param>
+	/// <param name="bccAddresses">BCC recipients.</param>
 	/// <param name="attachments">The attachments to an e-mail.</param>
-	/// <returns>Process status, <see langword="true"/> if all messages are processed to sent successfully</returns>
 	void Send(string fromMailAddress, IList<string> addresses, IList<string> ccAddresses, string subject, string body,
-		string bodyForAntiSpam = null, params MimeEntity[] attachments);
+		string bodyForAntiSpam = null, IList<string> bccAddresses = null, params MimeEntity[] attachments);
 
 	/// <summary>
 	/// Send e-mail to multiple recipients and carbon copy recipients in one e-mail asynchronously.
@@ -226,11 +218,11 @@ public interface IMailSender : IDisposable
 	/// <param name="subject">e-mail subject.</param>
 	/// <param name="body">e-mail body.</param>
 	/// <param name="bodyForAntiSpam">Part of an e-mail body just for anti-spam checking.</param>
+	/// <param name="bccAddresses">BCC recipients.</param>
 	/// <param name="attachments">The attachments to an e-mail.</param>
 	/// <param name="cancellationToken">The cancellation token.</param>
-	/// <returns>Process status, <see langword="true"/> if all messages are processed to sent successfully</returns>
 	Task SendAsync(string fromMailAddress, IList<string> addresses, IList<string> ccAddresses, string subject, string body,
-		string bodyForAntiSpam = null, MimeEntity[] attachments = null, CancellationToken cancellationToken = default);
+		string bodyForAntiSpam = null, IList<string> bccAddresses = null, MimeEntity[] attachments = null, CancellationToken cancellationToken = default);
 
 	/// <summary>
 	/// Send e-mail to multiple recipients separately
@@ -242,7 +234,6 @@ public interface IMailSender : IDisposable
 	/// <param name="body">e-mail body.</param>
 	/// <param name="bodyForAntiSpam">Part of an e-mail body just for anti-spam checking.</param>
 	/// <param name="attachments">The attachments to an e-mail.</param>
-	/// <returns>Process status, <see langword="true"/> if all messages are processed to sent successfully</returns>
 	void SendSeparately(SmtpClient client, string fromMailAddress, IList<string> addresses, string subject, string body,
 		string bodyForAntiSpam = null, params MimeEntity[] attachments);
 
@@ -257,7 +248,6 @@ public interface IMailSender : IDisposable
 	/// <param name="bodyForAntiSpam">Part of an e-mail body just for anti-spam checking.</param>
 	/// <param name="attachments">The attachments to an e-mail.</param>
 	/// <param name="cancellationToken">The cancellation token.</param>
-	/// <returns>Process status, <see langword="true"/> if all messages are processed to sent successfully</returns>
 	Task SendSeparatelyAsync(SmtpClient client, string fromMailAddress, IList<string> addresses, string subject, string body,
 		string bodyForAntiSpam = null, MimeEntity[] attachments = null, CancellationToken cancellationToken = default);
 
@@ -270,7 +260,6 @@ public interface IMailSender : IDisposable
 	/// <param name="body">e-mail body.</param>
 	/// <param name="bodyForAntiSpam">Part of an e-mail body just for anti-spam checking.</param>
 	/// <param name="attachments">The attachments to an e-mail.</param>
-	/// <returns>Process status, <see langword="true"/> if all messages are processed to sent successfully</returns>
 	void SendSeparately(string fromMailAddress, IList<string> addresses, string subject, string body,
 		string bodyForAntiSpam = null, params MimeEntity[] attachments);
 
@@ -284,7 +273,6 @@ public interface IMailSender : IDisposable
 	/// <param name="bodyForAntiSpam">Part of an e-mail body just for anti-spam checking.</param>
 	/// <param name="attachments">The attachments to an e-mail.</param>
 	/// <param name="cancellationToken">The cancellation token.</param>
-	/// <returns>Process status, <see langword="true"/> if all messages are processed to sent successfully</returns>
 	Task SendSeparatelyAsync(string fromMailAddress, IList<string> addresses, string subject, string body,
 		string bodyForAntiSpam = null, MimeEntity[] attachments = null, CancellationToken cancellationToken = default);
 }