+++ src/Simplify.Mail/MailSenderException.cs
@@ -5,12 +5,10 @@ namespace Simplify.Mail;
 /// <summary>
 /// The exception class using for MailSender exceptions
 /// </summary>
-[Serializable]
-public sealed class MailSenderException : Exception
+/// <remarks>
+/// Initializes a new instance of the <see cref="MailSenderException"/> class.
+/// </remarks>
+/// <param name="message">The message that describes the error.</param>
+public sealed class MailSenderException(string message) : Exception(message)
 {
-	/// <summary>
-	/// Initializes a new instance of the <see cref="MailSenderException"/> class.
-	/// </summary>
-	/// <param name="message">The message that describes the error.</param>
-	public MailSenderException(string message) : base(message) { }
 }